First, ugly pass at mnfstfwd
This monstronsity uses Ruby as the scripting language it wants to be.

Read a raw email in, parse out the sig and body, verify them, build a
JSON, POST it, and reply with the result. It even has error messages.

TODO: realize that I'm a programmer and not a sysadmin.